# Docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AUSTA Cockpit API
  description: |
    Medical Audit Platform API with AI-powered decision support.
    
    This API provides endpoints for managing medical authorization cases, 
    AI-assisted analysis, audit decisions, and comprehensive analytics.
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Obtain a token through the 
    `/auth/login` endpoint and include it in the Authorization header:
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API calls are rate limited to ensure fair usage:
    - Anonymous: 100 requests/hour
    - Authenticated: 1000 requests/hour
    - Premium: 10000 requests/hour
    
    ## Environments
    
    - Development: `http://localhost:3000/api/v1`
    - Staging: `https://staging-api.austa.com.br/v1`
    - Production: `https://api.austa.com.br/v1`
    
  version: 1.0.0
  contact:
    name: AUSTA Support
    email: suporte@austa.com.br
    url: https://docs.austa.com.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://staging-api.austa.com.br/v1
    description: Staging server
  - url: https://api.austa.com.br/v1
    description: Production server
...
paths:
...
  /ai/analyze/{caseId}:
    post:
      tags:
        - AI Analysis
      summary: Analyze case with AI
      description: Request AI analysis for a specific case
      operationId: analyzeCase
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CaseIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                analysisType:
                  type: string
                  enum: [full, medical, fraud, pattern]
                  default: full
                  description: Type of analysis to perform
                includeHistory:
                  type: boolean
                  default: true
                  description: Include patient history in analysis
                urgency:
                  type: string
                  enum: [normal, high]
                  default: normal
                  description: Analysis urgency
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAnalysis'
        '202':
          description: Analysis started, results will be available soon
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysisId:
                    type: string
                    format: uuid
                  estimatedTime:
                    type: integer
                    description: Estimated completion time in seconds
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /ai/chat:
    post:
      tags:
        - AI Analysis
      summary: Chat with AI assistant
      description: Send a message to the AI assistant for case-related questions
      operationId: chatWithAI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - caseId
              properties:
                message:
                  type: string
                  description: User's question or message
                  example: "What are the clinical guidelines for this procedure?"
                caseId:
                  type: string
                  format: uuid
                  description: Related case ID for context
                conversationId:
                  type: string
                  format: uuid
                  description: Optional conversation ID to continue a chat
      responses:
        '200':
          description: AI response generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIChatResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
    CaseIdParam:
      name: caseId
      in: path
      required: true
      description: Unique case identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AIAnalysis:
      type: object
      properties:
        id:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        analysisType:
          type: string
          enum: [full, medical, fraud, pattern]
        recommendation:
          type: string
          enum: [approve, deny, partial, review]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        medicalAnalysis:
          type: object
          properties:
            clinicalJustification:
              type: string
            guidelineCompliance:
              type: boolean
            similarCases:
              type: array
              items:
                type: object
                properties:
                  caseId:
                    type: string
                  similarity:
                    type: number
                  decision:
                    type: string
        fraudAnalysis:
          type: object
          properties:
            riskScore:
              type: number
              format: float
              minimum: 0
              maximum: 1
            riskFactors:
              type: array
              items:
                type: string
            anomalies:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  description:
                    type: string
                  severity:
                    type: string
                    enum: [low, medium, high]
        patternAnalysis:
          type: object
          properties:
            patterns:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  frequency:
                    type: integer
                  description:
                    type: string
        createdAt:
          type: string
          format: date-time
    AIChatResponse:
      type: object
      properties:
        conversationId:
          type: string
          format: uuid
        message:
          type: string
        confidence:
          type: number
          format: float
        sources:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [guideline, precedent, regulation, knowledge]
              reference:
                type: string
              relevance:
                type: number
                format: float
        suggestedQuestions:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            requestId:
              type: string
              format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per hour
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# Webhook Event Schemas
...